% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_transform.R
\name{bin_pseudotime}
\alias{bin_pseudotime}
\title{This function bins pseudotime values into a specified number of equal-width bins.}
\usage{
bin_pseudotime(x, n_bins = 100)
}
\arguments{
\item{x}{Numeric vector of pseudotime values}

\item{n_bins}{Number of bins to create (default: 100)}
}
\value{
An integer vector with bin assignments
}
\description{
This function bins pseudotime values into a specified number of equal-width bins.
}
