% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_transform.R
\name{calculate_bin_means}
\alias{calculate_bin_means}
\title{Calculate Bin Means Fast}
\usage{
calculate_bin_means(expression_matrix, bin_labels)
}
\arguments{
\item{expression_matrix}{A gene expression matrix (genes as rows, cells as columns)}

\item{bin_labels}{A vector of bin labels for each cell}
}
\value{
A matrix of mean expressions (genes as rows, bins as columns)
}
\description{
This function calculates the mean expression for each gene in each bin.
}
