% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{plot_results_brms}
\alias{plot_results_brms}
\title{Plot Results from Bayesian GAM Regression}
\usage{
plot_results_brms(fit)
}
\arguments{
\item{fit}{A model fit object returned by \code{bayesian_gam_regression_nb_shape}}
}
\value{
A grid arrangement of plots
}
\description{
This function generates diagnostic plots for the Bayesian GAM regression model.
}
