% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{prepare_data_for_gam}
\alias{prepare_data_for_gam}
\title{Prepare 3D Array Data for GAM Modeling}
\usage{
prepare_data_for_gam(gene_data)
}
\arguments{
\item{gene_data}{A 2D slice of a 3D array for a single gene [batch, time]}
}
\value{
A list containing:
  \item{x}{Vector of pseudotime values}
  \item{y}{Vector of expression values}
  \item{array_idx}{Vector of batch indices}
}
\description{
This function prepares data from a 3D array for GAM modeling.
}
