% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_multiple_models}
\alias{run_multiple_models}
\title{Run Trajectory Models for Multiple Genes}
\usage{
run_multiple_models(
  data_array,
  gene_indices = NULL,
  n_knots = 5,
  n_samples = 2000,
  parallel = FALSE,
  n_cores = 1
)
}
\arguments{
\item{data_array}{A 3D array of expression values [batch, time, gene]}

\item{gene_indices}{Indices of genes to model (default: all genes)}

\item{n_knots}{Number of knots for the cubic regression spline (default: 5)}

\item{n_samples}{Number of MCMC samples (default: 2000)}

\item{parallel}{Whether to run in parallel (default: FALSE)}

\item{n_cores}{Number of cores to use for parallel processing (default: 1)}
}
\value{
A list of fitted model objects
}
\description{
This function runs Bayesian GAM models for multiple genes across a trajectory.
}
