% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_trajectory_model}
\alias{run_trajectory_model}
\title{Run Bayesian GAM Model for a Single Gene}
\usage{
run_trajectory_model(
  data_array,
  gene_index,
  n_knots = 5,
  n_samples = 2000,
  backend = "cmdstanr"
)
}
\arguments{
\item{data_array}{A 3D array of expression values [batch, time, gene]}

\item{gene_index}{Index of the gene to model in the 3D array}

\item{n_knots}{Number of knots for the cubic regression spline (default: 5)}

\item{n_samples}{Number of MCMC samples (default: 2000)}

\item{backend}{Backend to use for Stan models ("rstan" or "cmdstanr", defaults to "rstan")}
}
\value{
A fitted model object from \code{bayesian_gam_regression_nb_shape}
}
\description{
This function runs a Bayesian GAM model for a single gene across a trajectory.
}
