% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajConserve-package.R
\docType{package}
\name{trajConserve}
\alias{TrajConserve}
\alias{TrajConserve-package}
\alias{trajConserve}
\alias{trajConserve-package}
\title{TrajConserve: Trajectory Conservation Analysis Tools for Single-Cell Data}
\description{
Tools for analyzing trajectory conservation in single-cell data. This package 
provides functions for Bayesian GAM regression modeling of expression trajectories, 
pseudotime binning and transformation, and visualization of trajectory models.
}
\section{Key Functions}{

\itemize{
  \item \code{\link{seurat_to_trajectory_array}} - Convert Seurat object to a 3D trajectory array
  \item \code{\link{bayesian_gam_regression_nb_shape}} - Fit Bayesian GAM regression model
  \item \code{\link{run_multiple_models}} - Run models for multiple genes in parallel
  \item \code{\link{plot_results_brms}} - Visualize model results
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GilbertHan1011/TrajConserve}
  \item Report bugs at \url{https://github.com/GilbertHan1011/TrajConserve/issues}
}

}
\author{
Gilbert Han
}
\keyword{package}
